datakit/dataviz: absorb viz store-cache sampling, remove viz/
the standalone experiments/datakit/viz/ overlapped dataviz almost entirely; its one unique capability was sampling the store's per-bucket tokenized Levanter caches directly and detokenizing (ground truth of what the store holds, vs dataviz's ducky reconstruction from upstream parquet).

- add `dataviz/store_cache.py` (`StoreCacheSampler`): lazy per-(cluster,quality) cache reads + contiguous-window sampling + detokenize
- register `store_cache_samples` view; add a 'from store cache' toggle to the Store tab
- delete `experiments/datakit/viz/` (server.py + index.html)
